refactor(cli): share yaml execution summary logic
